refactor the interactive module
This change deletes the old 'lex' module and creates a new module
called "Interactive". The move is in anticipation of having a bot
with two main fulfillment strategies, commands, and interactive
conversations.

Regardless of the strategy, a user's message will eventually be
resolved into a 'BotAction' which does the work of actually DOING
something.